add: tui package
