+++ client/internal/chat/status.go
@@ -1,0 +1,26 @@
+package chat
+
+import (
+	"fmt"
+)
+
+type ConnStatus uint64
+
+const (
+	Unknown ConnStatus = iota
+	Online
+	Offline
+)
+
+func (s ConnStatus) String() string {
+	switch s {
+	case Online:
+		return "Online"
+	case Offline:
+		return "Offline"
+	case Unknown:
+		return "Unknown"
+	default:
+		return fmt.Sprintf("ConnStatus-(%d)", s)
+	}
+}
