--- client/internal/tui/ascii-const.go
@@ -1,1 +1,0 @@
-package tui
